Update ppiclf_solve.f
I found this to be useful to allow 1 bin for the entire domain. This is useful for single-particle simulations. The default behavior was to split the domain into 2 bins in each direction, resulting in 1 particle and 7 ghost particles. This could lead to particles being sent to the wrong bins due to truncation errors, since the original particle is at the intersection of all bins, so it could simultaneously go to any of the bins.